Add unit tests for new checkpointing functionality
This commit introduces several unit tests for the LayerWiseOptimizer's common state dictionary and dynamic inference engine, enhancing coverage for checkpointing scenarios. New test cases have been added to validate the functionality of loading and saving checkpoints across different configurations, including layer-wise and hybrid models. Additionally, the test suite has been updated to include tests for dynamic inference with parallel configurations.
